Updated target to simplify circuit and make immune to spurious light sources.
The design has been reduced in size from ~40mm diameter to ~21mm diameter. The sensor was changed from a single-color (red) detector to a full-spectrum + white sensor since the old version would get tripped by white light (and camera flashes). The microcontroller was changed from the MSP430G2553 to the MSP430G2452 in the 14 pin package to save space. The new target will mount from the bottom and sit in the bottom of the table tennis ball instead of in the middle.