Extend BroadcastReset Delays to ensure they are detected on the other targets and add ifdef to disable pullups.
If pullups are enabled on all the targets then when a target detects a hit it will not be able to pull the line low. We could add pullups on the control board but since there is no regulation there it would require we add a regulator then pull the lines high. The simple solution is to enable internal pullups on one of the targets and just make sure none of the other targets have them enabled.